+++ src/app/containers/OutcomeSet/index.tsx
@@ -1,15 +1,13 @@
 import * as React from 'react';
 import { Helmet } from 'react-helmet';
-import {EditQuestionnaireName} from 'components/EditQuestionnaireName';
-import {EditQuestionnaireDescription} from 'components/EditQuestionnaireDescription';
-import {EditQuestionnaireInstructions} from 'components/EditQuestionnaireInstructions';
 import {IOutcomeResult, getOutcomeSet, IOutcomeMutation} from 'apollo/modules/outcomeSets';
-import {IQuestionMutation, deleteQuestion} from 'apollo/modules/questions';
-import { Button, Icon, Grid, Loader } from 'semantic-ui-react';
-import {CategoryList} from 'components/CategoryList';
-import {QuestionList} from 'components/QuestionList';
-import {Hint} from 'components/Hint';
-const strings = require('./../../../strings.json');
+import {IQuestionMutation} from 'apollo/modules/questions';
+import { Grid, Loader, Menu } from 'semantic-ui-react';
+import {SecondaryMenu} from 'components/SecondaryMenu';
+import { General } from './general';
+import { Questions } from './questions';
+import { Categories } from './categories';
+import { Switch, Route } from 'react-router-dom';
 import './style.less';
 
 interface IProps extends IOutcomeMutation, IQuestionMutation {
@@ -18,107 +16,50 @@ interface IProps extends IOutcomeMutation, IQuestionMutation {
     params: {
       id: string,
     },
+    path: string,
+    url: string,
   };
 }
 
-interface IState {
-  displayEditNameControl?: boolean;
-  displayEditDescriptionControl?: boolean;
-  editingInstructions?: boolean;
-}
-
-class OutcomeSetInner extends React.Component<IProps, IState> {
-  constructor(props) {
-    super(props);
-    this.state = {};
-    this.setEditNameState = this.setEditNameState.bind(this);
-    this.setEditDescState = this.setEditDescState.bind(this);
-    this.setEditInstructionsState = this.setEditInstructionsState.bind(this);
-    this.renderEditButton = this.renderEditButton.bind(this);
-  }
-
-  private setEditNameState(displayed: boolean): () => void {
-    return () => {
-      this.setState({
-        displayEditNameControl: displayed,
-      });
-    };
-  }
-
-  private setEditDescState(displayed: boolean): () => void {
-    return () => {
-      this.setState({
-        displayEditDescriptionControl: displayed,
-      });
-    };
-  }
-
-  private setEditInstructionsState(displayed: boolean): () => void {
-    return () => {
-      this.setState({
-        editingInstructions: displayed,
-      });
-    };
-  }
-
-  private renderEditButton(onClick: () => void): JSX.Element {
-    return (
-      <Button icon={true} basic={true} circular={true} size="mini" onClick={onClick}>
-        <Icon name="pencil"/>
-      </Button>
-    );
-  }
-
+class OutcomeSetInner extends React.Component<IProps, any> {
   public render() {
-    const wrapper = (inner: JSX.Element): JSX.Element => {
+    const wrapper = (inner: JSX.Element, signpost?: string): JSX.Element => {
       return (
-        <Grid container={true} columns={1} id="question-set">
-          <Grid.Column>
-            <Helmet>
-              <title>Questionnaire</title>
-            </Helmet>
-            <div>
-              {inner}
-            </div>
-          </Grid.Column>
-        </Grid>
+        <div>
+          <SecondaryMenu signpost={signpost}>
+            <Menu.Item name="General" />
+            <Menu.Item name="Questions" />
+            <Menu.Item name="Categories" />
+          </SecondaryMenu>
+          <Grid container={true} columns={1} id="question-set">
+            <Grid.Column>
+              <Helmet>
+                <title>Questionnaire</title>
+              </Helmet>
+              <div>
+                {inner}
+              </div>
+            </Grid.Column>
+          </Grid>
+        </div>
       );
     };
-    const { data, match: {params} } = this.props;
-    const { displayEditNameControl, displayEditDescriptionControl, editingInstructions } = this.state;
-    if (data.loading) {
-      return wrapper(<Loader active={true} inline="centered" />);
+    const { data: {loading, getOutcomeSet}} = this.props;
+    if (loading) {
+      return wrapper((<Loader active={true} inline="centered" />), 'Loading...');
     }
-    const os = data.getOutcomeSet;
-    if (os === undefined) {
-        return wrapper(<div />);
+    if (getOutcomeSet === undefined) {
+      return wrapper((<div />), 'Loading...');
     }
+    const match = this.props.match.path;
     return wrapper((
-      <div>
-        {displayEditNameControl ?
-          <EditQuestionnaireName data={data} outcomeSetID={params.id} afterSubmit={this.setEditNameState(false)} />
-          :
-          <h1>{os.name}{this.renderEditButton(this.setEditNameState(true))}</h1>
-        }
-        {displayEditDescriptionControl ?
-          <EditQuestionnaireDescription data={data} outcomeSetID={params.id} afterSubmit={this.setEditDescState(false)} />
-          :
-          <p>{os.description || 'No description'}{this.renderEditButton(this.setEditDescState(true))}</p>
-        }
-        <h3>Instructions <Hint text={strings.instructionsExplanation} /></h3>
-        {editingInstructions ?
-          <EditQuestionnaireInstructions data={data} outcomeSetID={params.id} afterSubmit={this.setEditInstructionsState(false)} />
-          :
-          <p className="instructions">{os.instructions || 'No instructions'}{this.renderEditButton(this.setEditInstructionsState(true))}</p>
-        }
-        <h3>Questions</h3>
-        <QuestionList outcomeSetID={params.id} data={this.props.data}/>
-        <h3>Question Categories <Hint text={strings.questionCategoryExplanation} /></h3>
-        <CategoryList outcomeSetID={params.id} data={this.props.data}/>
-      </div>
-    ));
+      <Switch>
+        <Route exact={true} path={`${match}/`} component={General} />
+        <Route path={`${match}/questions`} component={Questions} />
+        <Route path={`${match}/categories`} component={Categories} />
+      </Switch>
+    ), getOutcomeSet.name);
   }
 }
 
-const OutcomeSet = getOutcomeSet<IProps>((props) => props.match.params.id)(deleteQuestion<IProps>(OutcomeSetInner));
-export { OutcomeSet };
+export const OutcomeSet = getOutcomeSet<IProps>((props) => props.match.params.id)(OutcomeSetInner);
